feat: updated MFE structure and reduced re-rendering for sidebar
