+++ src/components/NavigationBar/CourseTabsNavigation.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import React, { memo, useEffect } from 'react';
 import PropTypes from 'prop-types';
 
 import classNames from 'classnames';
@@ -6,25 +6,30 @@ import { useDispatch, useSelector } from 'react-redux';
 
 import { useIntl } from '@edx/frontend-platform/i18n';
 
+import withConditionalInContextRendering from '../../discussions/common/withConditionalInContextRendering';
+import { useCourseId } from '../../discussions/data/hooks';
 import { fetchTab } from './data/thunks';
 import Tabs from './tabs/Tabs';
 import messages from './messages';
 
 import './navBar.scss';
 
-const CourseTabsNavigation = ({
-  activeTab, className, courseId, rootSlug,
-}) => {
+const CourseTabsNavigation = ({ activeTab, className, rootSlug }) => {
   const dispatch = useDispatch();
   const intl = useIntl();
+  const courseId = useCourseId();
   const tabs = useSelector(state => state.courseTabs.tabs);
 
   useEffect(() => {
-    dispatch(fetchTab(courseId, rootSlug));
+    if (courseId) {
+      dispatch(fetchTab(courseId, rootSlug));
+    }
   }, [courseId]);
 
+  console.log('CourseTabsNavigation');
+
   return (
-    <div id="courseTabsNavigation" className={classNames('course-tabs-navigation px-4', className)}>
+    <div id="courseTabsNavigation" tabIndex="-1" className={classNames('course-tabs-navigation px-4', className)}>
       {!!tabs.length && (
         <Tabs
           className="nav-underline-tabs"
@@ -49,13 +54,12 @@ CourseTabsNavigation.propTypes = {
   activeTab: PropTypes.string,
   className: PropTypes.string,
   rootSlug: PropTypes.string,
-  courseId: PropTypes.string.isRequired,
 };
 
 CourseTabsNavigation.defaultProps = {
-  activeTab: undefined,
+  activeTab: 'discussion',
   className: null,
   rootSlug: 'outline',
 };
 
-export default React.memo(CourseTabsNavigation);
+export default memo(withConditionalInContextRendering(CourseTabsNavigation, false));
