+++ src/discussions/navigation/navigation-bar/NavigationBar.jsx
@@ -1,4 +1,4 @@
-import React, { useContext, useMemo } from 'react';
+import React, { memo, useMemo } from 'react';
 
 import { matchPath } from 'react-router';
 import { NavLink } from 'react-router-dom';
@@ -7,14 +7,14 @@ import { useIntl } from '@edx/frontend-platform/i18n';
 import { Nav } from '@edx/paragon';
 
 import { Routes } from '../../../data/constants';
-import { DiscussionContext } from '../../common/context';
-import { useShowLearnersTab } from '../../data/hooks';
+import withConditionalInContextRendering from '../../common/withConditionalInContextRendering';
+import { useCourseId, useShowLearnersTab } from '../../data/hooks';
 import { discussionsPath } from '../../utils';
 import messages from './messages';
 
 const NavigationBar = () => {
   const intl = useIntl();
-  const { courseId } = useContext(DiscussionContext);
+  const courseId = useCourseId();
   const showLearnersTab = useShowLearnersTab();
 
   const navLinks = useMemo(() => ([
@@ -41,23 +41,28 @@ const NavigationBar = () => {
       });
     }
   }, [showLearnersTab]);
+  console.log('NavigationBar');
+
+  const navLinksList = useMemo(() => (
+    navLinks.map(link => (
+      <Nav.Item key={link.route}>
+        <Nav.Link
+          key={link.route}
+          as={NavLink}
+          to={discussionsPath(link.route, { courseId })}
+          isActive={link.isActive}
+        >
+          {intl.formatMessage(link.labelMessage)}
+        </Nav.Link>
+      </Nav.Item>
+    ))
+  ), [navLinks]);
 
   return (
     <Nav variant="pills" className="py-2 nav-button-group">
-      {navLinks.map(link => (
-        <Nav.Item key={link.route}>
-          <Nav.Link
-            key={link.route}
-            as={NavLink}
-            to={discussionsPath(link.route, { courseId })}
-            isActive={link.isActive}
-          >
-            {intl.formatMessage(link.labelMessage)}
-          </Nav.Link>
-        </Nav.Item>
-      ))}
+      {navLinksList}
     </Nav>
   );
 };
 
-export default React.memo(NavigationBar);
+export default memo(withConditionalInContextRendering(NavigationBar, false));
