+++ src/discussions/tours/DiscussionsProductTour.jsx
@@ -1,31 +1,23 @@
-import React, { useEffect } from 'react';
+import React, { memo } from 'react';
 
 import isEmpty from 'lodash/isEmpty';
-import { useDispatch } from 'react-redux';
 
 import { ProductTour } from '@edx/paragon';
 
+import withConditionalInContextRendering from '../common/withConditionalInContextRendering';
 import { useTourConfiguration } from '../data/hooks';
-import { fetchDiscussionTours } from './data/thunks';
 
 const DiscussionsProductTour = () => {
-  const dispatch = useDispatch();
   const config = useTourConfiguration();
-
-  useEffect(() => {
-    dispatch(fetchDiscussionTours());
-  }, []);
+  console.log('DiscussionsProductTour');
 
   return (
-    // eslint-disable-next-line react/jsx-no-useless-fragment
-    <>
-      {!isEmpty(config) && (
-        <ProductTour
-          tours={config}
-        />
-      )}
-    </>
+    !isEmpty(config) && (
+      <ProductTour
+        tours={config}
+      />
+    )
   );
 };
 
-export default DiscussionsProductTour;
+export default memo(withConditionalInContextRendering(DiscussionsProductTour, false));
